refactor: presentation 외 다른 구조도 화면 별로 구성 - Config 클래스 제거 -> 변수로 변경
