+++ app/src/main/java/com/survivalcoding/stopwatch/Config.kt
@@ -1,9 +1,4 @@
 package com.survivalcoding.stopwatch
 
-class Config {
-    companion object {
-        // room
-        const val DATABASE_NAME = "stop_watch_db"
-        const val TABLE_NAME = "lap_time_table"
-    }
-}
+const val DATABASE_NAME = "stop_watch_db"
+const val TABLE_NAME = "lap_time_table"
