+++ app/src/main/java/com/survivalcoding/stopwatch/domain/stopwatch/repository/LapTimeRepository.kt
@@ -1,6 +1,6 @@
-package com.survivalcoding.stopwatch.domain.repository
+package com.survivalcoding.stopwatch.domain.stopwatch.repository
 
-import com.survivalcoding.stopwatch.domain.model.LapTimeRecord
+import com.survivalcoding.stopwatch.domain.stopwatch.model.LapTimeRecord
 import kotlinx.coroutines.flow.Flow
 
 interface LapTimeRepository {
