+++ app/src/main/java/com/survivalcoding/stopwatch/domain/stopwatch/use_case/DeleteAllLapTimesUseCase.kt
@@ -1,6 +1,6 @@
-package com.survivalcoding.stopwatch.domain.use_case
+package com.survivalcoding.stopwatch.domain.stopwatch.use_case
 
-import com.survivalcoding.stopwatch.domain.repository.LapTimeRepository
+import com.survivalcoding.stopwatch.domain.stopwatch.repository.LapTimeRepository
 
 class DeleteAllLapTimesUseCase(private val repository: LapTimeRepository) {
     suspend operator fun invoke() = repository.deleteAllLapTimes()
