--- app/src/main/java/com/survivalcoding/stopwatch/domain/use_case/InsertLapTimeUseCase.kt
@@ -1,9 +1,0 @@
-package com.survivalcoding.stopwatch.domain.use_case
-
-import com.survivalcoding.stopwatch.domain.model.LapTimeRecord
-import com.survivalcoding.stopwatch.domain.repository.LapTimeRepository
-
-class InsertLapTimeUseCase(private val repository: LapTimeRepository) {
-    suspend operator fun invoke(lapTimeRecord: LapTimeRecord) =
-        repository.insertLapTime(lapTimeRecord)
-}
